* Valores da variável tick para 64bit. - Isso corrige o problema de temporizadores parar de trabalhar depois de aproximadamente 24 a 50 horas * Correção ao banir personagem. * Adicionado suporte a opção de Game Master Remover equipamentos. * Correção em dano elemental. * Correção nas constantes de tamanho de jogador e monstros. * Adicionado Max EquipLv para modo Pré-Renovação e Old Times. * Correção de parênteses na condição status_calc_regen_rate.

+++ brathena/src/char/int_auction.c
@@ -34,7 +34,7 @@
 static DBMap *auction_db_ = NULL; // int auction_id -> struct auction_data*
 
 void auction_delete(struct auction_data *auction);
-static int auction_end_timer(int tid, unsigned int tick, int id, intptr_t data);
+static int auction_end_timer(int tid, int64 tick, int id, intptr_t data);
 
 static int auction_count(int char_id, bool buy)
 {
@@ -115,7 +115,7 @@ unsigned int auction_create(struct auction_data *auction)
 		auction->auction_id = 0;
 	} else {
 		struct auction_data *auction_;
-		unsigned int tick = auction->hours * 3600000;
+		int64 tick = auction->hours * 3600000;
 
 		auction->item.amount = 1;
 		auction->item.identify = 1;
@@ -146,7 +146,7 @@ static void mapif_Auction_message(int char_id, unsigned char result)
 	mapif_sendall(buf,7);
 }
 
-static int auction_end_timer(int tid, unsigned int tick, int id, intptr_t data)
+static int auction_end_timer(int tid, int64 tick, int id, intptr_t data)
 {
 	struct auction_data *auction;
 	if((auction = (struct auction_data *)idb_get(auction_db_, id)) != NULL) {
@@ -186,7 +186,7 @@ void inter_auctions_fromsql(void)
 	struct item *item;
 	char *data;
 	StringBuf buf;
-	unsigned int tick = gettick(), endtick;
+	int64 tick = gettick(), endtick;
 	time_t now = time(NULL);
 
 	StringBuf_Init(&buf);
@@ -232,7 +232,7 @@ void inter_auctions_fromsql(void)
 		}
 
 		if(auction->timestamp > now)
-			endtick = ((unsigned int)(auction->timestamp - now) * 1000) + tick;
+			endtick = ((int64)(auction->timestamp - now) * 1000) + tick;
 		else
 			endtick = tick + 10000; // 10 Second's to process ended auctions
