+++ brathena/src/map/battleground.h
@@ -110,9 +110,9 @@ struct battleground_interface {
 	int (*id2pos) (int queue_id, int account_id);
 	void (*queue_pc_cleanup) (struct map_session_data *sd);
 	void (*begin) (struct bg_arena *arena);
-	int (*begin_timer) (int tid, unsigned int tick, int id, intptr_t data);
+	int (*begin_timer) (int tid, int64 tick, int id, intptr_t data);
 	void (*queue_pregame) (struct bg_arena *arena);
-	int (*fillup_timer) (int tid, unsigned int tick, int id, intptr_t data);
+	int (*fillup_timer) (int tid, int64 tick, int id, intptr_t data);
 	void (*queue_ready_ack) (struct bg_arena *arena, struct map_session_data *sd, bool response);
 	void (*match_over) (struct bg_arena *arena, bool canceled);
 	void (*queue_check) (struct bg_arena *arena);
