+++ brathena/src/map/mapreg_sql.c
@@ -188,7 +188,7 @@ static void script_save_mapreg(void)
 	mapreg_dirty = false;
 }
 
-static int script_autosave_mapreg(int tid, unsigned int tick, int id, intptr_t data)
+static int script_autosave_mapreg(int tid, int64 tick, int id, intptr_t data)
 {
 	if(mapreg_dirty)
 		script_save_mapreg();
